Reading and sending real soil moisture works!
