+++ code/nrf-connect/samples/zigbee/prj.conf
@@ -6,9 +6,20 @@
 
 CONFIG_NCS_SAMPLES_DEFAULTS=y
 
+CONFIG_CBPRINTF_FP_SUPPORT=y
+CONFIG_PWM=y
+CONFIG_I2C=y
+CONFIG_ADC=y
+CONFIG_GPIO=y
+
+CONFIG_PM=y
+CONFIG_PM_DEVICE=y
+
+CONFIG_NEWLIB_LIBC=y
+CONFIG_NEWLIB_LIBC_FLOAT_PRINTF=y
+
 CONFIG_UART_INTERRUPT_DRIVEN=y
 CONFIG_SERIAL=y
-CONFIG_GPIO=y
 
 # Make sure printk is not printing to the UART console
 CONFIG_CONSOLE=y
